DAO support for loading Attributes in bulk

+++ muck-service/src/main/kotlin/uk/co/grahamcox/muck/service/model/Page.kt
@@ -5,6 +5,6 @@ package uk.co.grahamcox.muck.service.model
  */
 data class Page<out DATA>(
         val data: List<DATA>,
-        val offset: Long,
-        val total: Long
+        val offset: Int,
+        val total: Int
 )
